Allow user configuration of endpoints
This exposes the end points described in the metadata.xml file
generated when calling Net::SAML2::SP’s ->metadata for the
XML file.

The following attributes can now be set through Net::SAML2::SP->new,
and must be set with the full URL of the endpoint:

* soap
* http_redirect
* http_post
* http_artifact

When a value is not set, it defaults to the URL Net::SAML2 automatically
set, for backwards compatibility.